create DirectoryListItem component
